+++ chat-components/src/components/citationpane/interfaces/ICitationPaneComponentOverrides.ts
@@ -1,0 +1,8 @@
+import { ReactNode } from "react";
+
+export interface ICitationPaneComponentOverrides {
+    title?: ReactNode | string;
+    subtitle?: ReactNode | string;
+    closeButton?: ReactNode | string;
+    topCloseButton?: ReactNode | string;
+}